# Micro optimization on 'gc_invalidateOrderedFinalizerList'