@undingen 2016-01-04
I'm was a little bit undecided about this change at first because the perf changes are just noise (I could not reproduce it) and looking at the stats output I can only see a very minimal change :-(. It does reduce the ``` us_gc_invalidate_ordered_finalizer_list``` number a lot but this is already very fast compared to the time we spend in other parts of the GC so it does not show up when running a benchmark.
But as the small change is always a very tiny improvement and does not make the code more complicated I decided to merge it in.

Thanks very much for the patch!